+++ src/jestutils/jestHelpers.ts
@@ -46,6 +46,13 @@ import { MATTER_STORAGE_NAME, Matterbridge, MatterbridgePlatform } from 'matterb
 */
 
 export let loggerLogSpy: jest.SpiedFunction<typeof AnsiLogger.prototype.log>;
+export let loggerDebugSpy: jest.SpiedFunction<typeof AnsiLogger.prototype.debug>;
+export let loggerInfoSpy: jest.SpiedFunction<typeof AnsiLogger.prototype.info>;
+export let loggerNoticeSpy: jest.SpiedFunction<typeof AnsiLogger.prototype.notice>;
+export let loggerWarnSpy: jest.SpiedFunction<typeof AnsiLogger.prototype.warn>;
+export let loggerErrorSpy: jest.SpiedFunction<typeof AnsiLogger.prototype.error>;
+export let loggerFatalSpy: jest.SpiedFunction<typeof AnsiLogger.prototype.fatal>;
+
 export let consoleLogSpy: jest.SpiedFunction<typeof console.log>;
 export let consoleDebugSpy: jest.SpiedFunction<typeof console.log>;
 export let consoleInfoSpy: jest.SpiedFunction<typeof console.log>;
@@ -88,6 +95,12 @@ export async function setupTest(name: string, debug: boolean = false): Promise<v
   rmSync(path.join('jest', name), { recursive: true, force: true });
 
   const { jest } = await import('@jest/globals');
+  loggerDebugSpy = jest.spyOn(AnsiLogger.prototype, 'debug');
+  loggerInfoSpy = jest.spyOn(AnsiLogger.prototype, 'info');
+  loggerNoticeSpy = jest.spyOn(AnsiLogger.prototype, 'notice');
+  loggerWarnSpy = jest.spyOn(AnsiLogger.prototype, 'warn');
+  loggerErrorSpy = jest.spyOn(AnsiLogger.prototype, 'error');
+  loggerFatalSpy = jest.spyOn(AnsiLogger.prototype, 'fatal');
   if (debug) {
     loggerLogSpy = jest.spyOn(AnsiLogger.prototype, 'log');
     consoleLogSpy = jest.spyOn(console, 'log');